+++ android/app/src/main/java/com/shahartal/myquietchannel/parasha/HebCalZmanimModel.kt
@@ -1,5 +1,6 @@
 package com.shahartal.myquietchannel.parasha
 
 data class HebCalZmanimModel(
+    val location: HebCalLocationModel,
     val times: HebCalZmanimTimesModel
 )